app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
lxml==5.2.2
anthropic==0.34.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.0
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==24.1.0
httpx==0.27.2
itsdangerous==2.2.0
openpyxl==3.1.5
